django/backend/requirements.txt
django
psycopg-binary
psycopg
djangorestframework
djangorestframework_simplejwt
django-cors-headers
daphne
requests
channels
pyotp
pillow